modify some bug

+++ models/cogvideox/pipeline_cogvideox.py
@@ -143,7 +143,7 @@ class CogVideoXSGPipeline(CogVideoXPipeline):
         self_guidance_scale: float = 1.0,
         self_guidance_shift_t: float = 10,
         self_guidance_type: str = "normal",
-        sg_prev_max_t: int = 500,
+        sg_prev_max_t: float = 500,
     ) -> Union[CogVideoXPipelineOutput, Tuple]:
         """
         Function invoked when calling the pipeline for generation.
