+++ models/stable_diffusion/pipeline_sd.py
@@ -138,7 +138,7 @@ class StableDiffusionSGPAGPipeline(StableDiffusionPAGPipeline):
         self_guidance_scale: float = 1.0,
         self_guidance_shift_t: float = 10,
         self_guidance_type: str = "normal",
-        sg_prev_max_t: int = 500,
+        sg_prev_max_t: float = 500,
     ):
         r"""
         The call function to the pipeline for generation.
